requirements-dev.txt
-r requirements.txt

black==22.3.0
coverage==6.3.3
factory-boy==3.2.1
Faker==13.7.0
mypy==0.950
mypy-extensions==0.4.3
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-factoryboy==2.2.0
pathspec==0.9.0
platformdirs==2.5.2
types-requests==2.27.27
types-urllib3==1.26.14
